Fix Nodemailer on netlify
Add the required API keys to netlify but I needed to move back our
api endpoint to `/src/pages/api`. I had blindly listened to ChatGPT to
move it into `/netlify/functions` but I forgot that the astro netlify
adapter is supposed to handle that. So if we deploy to netlify, we
definitely need the adapter. But we can deploy as `static` vs `server`
because we only have one componenet that is not static `contact.ts`. For
ths to work, nonstatic functions like contact.ts need to
`export prerender = false`.

eslint should avoid the `.netlify` folder.

Modify `Form.astro` to prevent clicking the submit button
until a response from the server. This is because it was possible to
click submit very quickly in succession to send duplicate emails as many
times as clicked before the server responded and form reset.

Add npm run deploy for deploying to netlify